Code is cleaned up
