+++ user_server/service/service.go
@@ -34,3 +34,25 @@ func (s *UserService) CreateUser(ctx context.Context, user *user.User) (*user.Us
 
 	return user, nil
 }
+
+func (s *UserService) ReadUser(ctx context.Context, req *user.UserId) (*user.User, error) {
+	if req == nil || req.Id < 0 {
+		return nil, fmt.Errorf("invalid id %v", req.Id)
+	}
+
+	query := "SELECT * FROM users WHERE id = ?"
+
+	row := s.db.QueryRow(query, req.Id)
+
+	var user user.User
+
+	// scan the row in the user struct
+
+	err := row.Scan(&user.Id, &user.Name, &user.Email, &user.Password, &user.Role, &user.IsActive, &user.CreatedAt, &user.UpdatedAt)
+	if err != nil {
+		return nil, err
+	}
+
+	return &user, nil
+
+}
